APM-1112 moved log history component to app level and added logs for updating process and connector
