--- projects/ama-sdk/src/lib/helpers/components/editor-footer/editor-footer.component.ts
@@ -1,39 +1,0 @@
- /*!
- * @license
- * Copyright 2019 Alfresco, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { Component, Inject } from '@angular/core';
-import { EDITOR_FOOTER_SERVICE_TOKEN, EditorFooterService } from '../../../services/editor-footer.service.interface';
-
-@Component({
-    selector: 'amasdk-editor-footer',
-    templateUrl: './editor-footer.component.html'
-})
-export class EditorFooterComponent {
-
-    showConsole = false;
-
-    constructor(@Inject(EDITOR_FOOTER_SERVICE_TOKEN) public editorFooterService: EditorFooterService) {}
-
-    toggleConsole() {
-        this.showConsole = !this.showConsole;
-        this.editorFooterService.setHistoryVisibility(this.showConsole);
-    }
-
-    clearLogs() {
-        this.editorFooterService.clearLogs();
-    }
-}
