+++ projects/ama-sdk/src/lib/services/public_api.ts
@@ -22,8 +22,8 @@
 export * from './uuid.service';
 export * from './title.service';
 export * from './download-resource.service';
-export * from './editor-footer.service.interface';
 export * from './ama-authentication.service';
 export * from './ama-services.module';
 export * from './auth.service';
 export * from './blob.service';
+export * from './editor-helper.service';
