+++ projects/ama-sdk/src/lib/store/public_api.ts
@@ -32,4 +32,5 @@ export * from './project.selectors';
 export * from './project-data.state';
 export * from './project-tree.state';
 export * from './project.actions';
+export * from './logging.actions';
