+++ src/app/app.module.ts
@@ -49,6 +49,9 @@ import { AmaRoleGuard } from './ama-role-guard.service';
 import { AppExtensionsModule } from './extensions.module';
 import { ModelStorageService } from './common/services/model-storage.service';
 import { GlobalErrorHandler } from './common/helpers/services/error-handler.service';
+import { LogHistoryComponent } from './app/app-layout/logging/components/log-history/log-history.component';
+import { LogHistoryEntryComponent } from './app/app-layout/logging/components/log-history/log-history-entry/log-history-entry.component';
+import { EditorFooterComponent } from './app/app-layout/editor-footer/editor-footer.component';
 
 @NgModule({
     imports: [
@@ -89,7 +92,10 @@ import { GlobalErrorHandler } from './common/helpers/services/error-handler.serv
         ErrorContentComponent,
         CurrentUserComponent,
         SettingsDialogComponent,
-        HostSettingsComponent
+        HostSettingsComponent,
+        LogHistoryComponent,
+        LogHistoryEntryComponent,
+        EditorFooterComponent
     ],
     entryComponents: [SettingsDialogComponent],
     providers: [
