+++ src/app/app/app-layout/app-layout.component.html
@@ -30,7 +30,7 @@
     <adf-sidenav-layout-content>
         <ng-template>
             <router-outlet></router-outlet>
-            <amasdk-editor-footer class="editor-footer"></amasdk-editor-footer>
+            <ama-editor-footer class="editor-footer">TEST</ama-editor-footer>
         </ng-template>
     </adf-sidenav-layout-content>
