+++ src/app/app/app-layout/editor-footer/editor-footer.component.html
@@ -10,7 +10,7 @@
             </button>
         </div>
     </mat-toolbar>
-    <amasdk-log-history class="log-history" [messages]="editorFooterService.logs$ | async"></amasdk-log-history>
+    <ama-log-history class="log-history" [messages]="editorFooterService.logs$ | async"></ama-log-history>
 </ng-container>
 <mat-toolbar class="editor-footer-toolbar">
     <div class="editor-footer-toolbar-left">
