+++ src/app/app/app-layout/editor-footer/editor-footer.service.interface.ts
@@ -17,7 +17,7 @@
 
 import { InjectionToken } from '@angular/core';
 import { Observable } from 'rxjs';
-import { LogMessage } from '../store/app.state';
+import { LogMessage } from 'ama-sdk';
 
 export interface EditorFooterService {
     userMessage$: Observable<string>;
