+++ src/app/app/app-layout/logging/components/log-history/log-history-entry/log-history-entry.component.html
