+++ src/app/app/app-layout/logging/components/log-history/log-history.component.html
@@ -1,5 +1,5 @@
 <section class="log-history-panel">
     <ng-container *ngFor="let log of messages">
-        <amasdk-log-history-entry [log]="log"></amasdk-log-history-entry>
+        <ama-log-history-entry [log]="log"></ama-log-history-entry>
     </ng-container>
 </section>
