+++ src/app/app/app-layout/logging/components/log-history/log-history.component.scss
