+++ src/app/common/services/app-footer.service.ts
@@ -19,11 +19,12 @@
 import { Injectable } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { Observable, combineLatest, BehaviorSubject } from 'rxjs';
-import { EditorFooterService, LogMessage, MESSAGE } from 'ama-sdk';
+import { LogMessage, MESSAGE } from 'ama-sdk';
 import { map, tap } from 'rxjs/operators';
 import { ClearLogHistoryAction, SetLogHistoryVisibilityAction } from '../../../app/store/actions/app.actions';
 import { selectToolbarUserMessage, selectToolbarInProgress, selectToolbarLogs, selectToolbarLogsVisibility } from '../../../app/store/selectors/app.selectors';
 import { AmaState } from 'ama-sdk';
+import { EditorFooterService } from './../../app/app-layout/editor-footer/editor-footer.service.interface';
 
 @Injectable()
 export class AppFooterService implements EditorFooterService  {
