+++ src/app/process-editor/components/process-editor/process-editor.component.html
@@ -7,7 +7,7 @@
                 [breadcrumbs$]="breadcrumbs$"
                 [content]="content$ | async">
             </ama-process-header>
-            <mat-tab-group [@.disabled]="true" class="ama-process-tab-group" *ngIf="process$ | async as process;" (selectedTabChange)="selectedTabChange($event)">
+            <mat-tab-group [@.disabled]="true" class="ama-process-tab-group" *ngIf="process$ | async as process;">
                 <mat-tab label="{{ 'PROCESS_EDITOR.TABS.DIAGRAM_EDITOR' | translate }}">
                     <div class="process-editor-modeling-area" cdkDropList>
                         <ama-process-modeler
@@ -25,7 +25,6 @@
                     <div class="ama-process-code-editor" data-automation-id="process-basic-tab">
                         <amasdk-code-editor
                             (changed)="onXmlChangeAttempt($event)"
-                            (positionChanged)="codeEditorPositionChanged($event)"
                             [vsTheme]="vsTheme$ | async"
                             [options]="{language: 'xml'}"
                             [content]="content$ | async"
@@ -38,7 +37,6 @@
                     <div class="ama-process-code-editor" data-automation-id="extensions-tab">
                         <amasdk-code-editor
                             (changed)="onExtensionsChangeAttempt($event, process.id)"
-                            (positionChanged)="codeEditorPositionChanged($event)"
                             [vsTheme]="vsTheme$ | async"
                             [content]="extensions$ | async"
                             [fileUri]="extensionFileUri$ | async"
